+++ wp-admin/includes/class-wp-plugin-dependencies.php
@@ -332,10 +332,50 @@ class WP_Plugin_Dependencies {
 	public function modify_plugin_row_elements_requires( $plugin_file ) {
 		$names = $this->get_requires_plugins_names( $plugin_file );
 		if ( ! empty( $names ) ) {
+			$names = $this->get_view_details_link( $plugin_file, $names );
 			print '<script>';
-			print 'jQuery("tr[data-plugin=\'' . esc_attr( $plugin_file ) . '\'] .plugin-version-author-uri").append("<br><br><strong>' . esc_html__( 'Requires:' ) . '</strong> ' . esc_html( $names ) . '");';
+			print 'jQuery("tr[data-plugin=\'' . esc_attr( $plugin_file ) . '\'] .plugin-version-author-uri").append("<br><br><strong>' . esc_html__( 'Requires:' ) . '</strong> ' . wp_kses_post( $names ) . '");';
 			print '</script>';
 		}
+	}
+
+	/**
+	 * Create 'View details' like links for required plugins.
+	 *
+	 * @param string $plugin_file Plugin file name.
+	 * @param string $names       Names of required plugins.
+	 *
+	 * @return string
+	 */
+	private function get_view_details_link( $plugin_file, $names ) {
+		$details_links = array();
+		$names_arr     = explode( ', ', $names );
+		$dependencies  = $this->requires_plugins[ $plugin_file ]['RequiresPlugins'];
+		$dependencies  = explode( ',', $dependencies );
+
+		foreach ( $dependencies as $dependency ) {
+			$plugin_data = $this->plugin_data[ $dependency ];
+			foreach ( $names_arr as $name ) {
+				if ( $name === $plugin_data['name'] ) {
+					if ( empty( $plugin_data['version'] ) ) {
+						$details_links[ $name ] = $name;
+					} else {
+						$details_links[ $name ] = sprintf(
+							"<a href='%s' class='thickbox open-plugin-details-modal' aria-label='%s' data-title='%s'>%s</a>",
+							network_admin_url( 'plugin-install.php?tab=plugin-information&plugin=' . $plugin_data['slug'] . '&TB_iframe=true&width=600&height=550' ),
+							/* translators: %s: Plugin name. */
+							sprintf( __( 'More information about %s' ), $name ),
+							$name,
+							$name
+						);
+					}
+				}
+			}
+		}
+
+		$details_links = implode( ', ', $details_links );
+
+		return $details_links;
 	}
 
 	/**