+++ CHANGES.md
@@ -1,6 +1,7 @@
 [unreleased]
 * change 'Activate' plugin action link to 'Cannot Activate' text when plugin has unmet dependencies
 * remove checkbox from plugin row when plugin has unmet dependencies
+* use _View details_ link for plugins listed in **Requires:** in plugin row
 
 #### 1.0.0 / 2022-05-31 🎂
 * fix typo